refactor: 라이브탭 컨테이너에서 사용하는 컨테이너 파일 분리
라이브탭 컨테이너에서 사용하는 LiveOn, LiveOff 컨테이너를 파일로 분리했습니다.
두 컴포넌트의 스타일이 중복되서 컨테이너/프로젝트/라이브탭 안에 작성했습니다.
